// packages/core/types/ExceptionCode.ts
export enum ExceptionCode {
  HO100100 = "HO100100",
  HO100101 = "HO100101",
  HO100102 = "HO100102",
  HO100103 = "HO100103",
  HO100106 = "HO100106",
  HO100107 = "HO100107",
  HO100108 = "HO100108",
  HO100109 = "HO100109",
  HO100110 = "HO100110",
  HO100200 = "HO100200",
  HO100201 = "HO100201",
  HO100202 = "HO100202",
  HO100203 = "HO100203",
  HO100206 = "HO100206",
  HO100207 = "HO100207",
  HO100208 = "HO100208",
  HO100209 = "HO100209",
  HO100210 = "HO100210",
  HO100211 = "HO100211",
  HO100212 = "HO100212",
  HO100213 = "HO100213",
  HO100215 = "HO100215",
  HO100216 = "HO100216",
  HO100217 = "HO100217",
  HO100218 = "HO100218",
  HO100219 = "HO100219",
  HO100220 = "HO100220",
  HO100228 = "HO100228",
  HO100232 = "HO100232",
  HO100233 = "HO100233",
  HO100234 = "HO100234",
  HO100236 = "HO100236",
  HO100237 = "HO100237",
  HO100239 = "HO100239",
  HO100240 = "HO100240",
  HO100241 = "HO100241",
  HO100242 = "HO100242",
  HO100243 = "HO100243",
  HO100244 = "HO100244",
  HO100245 = "HO100245",
  HO100246 = "HO100246",
  HO100247 = "HO100247",
  HO100249 = "HO100249",
  HO100300 = "HO100300",
  HO100301 = "HO100301",
  HO100302 = "HO100302",
  HO100304 = "HO100304",
  HO100305 = "HO100305",
  HO100306 = "HO100306",
  HO100307 = "HO100307",
  HO100309 = "HO100309",
  HO100310 = "HO100310",
  HO100312 = "HO100312",
  HO100311 = "HO100311",
  HO100313 = "HO100313",
  HO100314 = "HO100314",
  HO100315 = "HO100315",
  HO100320 = "HO100320",
  HO100321 = "HO100321",
  HO100322 = "HO100322",
  HO100323 = "HO100323",
  HO100324 = "HO100324",
  HO100325 = "HO100325",
  HO100326 = "HO100326",
  HO100328 = "HO100328",
  HO100329 = "HO100329",
  HO100331 = "HO100331",
  HO100332 = "HO100332",
  HO100333 = "HO100333",
  HO100402 = "HO100402",
  HO100404 = "HO100404",
  HO100507 = "HO100507",
  HO200101 = "HO200101",
  HO200104 = "HO200104",
  HO200108 = "HO200108",
  HO200113 = "HO200113",
  HO200114 = "HO200114",
  HO200100 = "HO200100",
  HO200103 = "HO200103",
  HO200106 = "HO200106",
  HO200107 = "HO200107",
  HO200109 = "HO200109",
  HO200110 = "HO200110",
  HO200111 = "HO200111",
  HO200112 = "HO200112",
  HO200115 = "HO200115",
  HO200116 = "HO200116",
  HO200117 = "HO200117",
  HO200118 = "HO200118",
  HO200119 = "HO200119",
  HO200120 = "HO200120",
  HO200121 = "HO200121",
  HO200122 = "HO200122",
  HO200123 = "HO200123",
  HO200124 = "HO200124",
  HO200200 = "HO200200",
  HO200201 = "HO200201",
  HO200202 = "HO200202",
  HO200203 = "HO200203",
  HO200205 = "HO200205",
  HO200209 = "HO200209",
  HO200210 = "HO200210",
  HO200211 = "HO200211",
  HO200212 = "HO200212"
}